[hotfix] Create JarUtils with common jar file tools
